Separate options to include bitmap terminals and old dot-matrix printers
One of the original reasons to disable bitmap terminals by default
was bitmap.c's license and that these terminals had seriously fallen
behind or just weren't useful anymore. Since the license changed, and the
bitmap terminals block and sixeltek can be useful, we only enable the
non-printer terminals with the "--with-bitmap-terminals" option.
This new option "--with-old-printers" additionally includes the terminals
for old dot-matrix printers.