Simplify contourfill by removing support for tic levels other than 0
I decided it was more confusing than helpful to allow contourfill
boundaries to be specified by setting custom level tics on z or cb.
If nothing else it leads to a conflict when a labelled major tic is
present at the same z-value.  The only code removed is set/show/save.
The actual test against tic level is a single line of code that still
exists to test against 0 (major tics only).