Try to recover from parsing errors during program initialization
Very early in program initialization the code tries to set a starting terminal
type by constructing a command based on the environment variable GNUTERM.
Essentially this is "set terminal ${GNUTERM}".  However at this stage
the program state is not yet fully initialized, so if a parsing error
or other problem is encountered the LONGJMP from int_error() restores an
incomplete state.

Now we try to detect this following the LONGJMP and resume execution
of the usual program initialization routines after setting the terminal
to "unknown".

Bug #2652