new command:  warn "message"
Trigger int_warn() from the command line.
This allows the user to print a message tagged with the script name
and the line that it came from.
As with other warnings, it is suppressed by "unset warnings".

This works also from inside a function block.
See discussion attached to Bug #2655