revert "using cbticlabels(N)"
This option, marked EXPERIMENTAL, was added for completeness since we
already had xticlabels, yticlabels, and so on.  But its usability is dubious.
It is at best ambiguous what coordinate the label would use.
If someone comes up with an actual use-case the requires this,
we can add it back or come up with an alternative.