plot/splot keyentry "sampletext" title "titletext"
"keyentry" followed directly by a text string places that text in the
position of the key sample.  The title text, if any, is placed in the
usual title text position.  The full syntax is

    keyentry {"sampletext" {<label options>}} title "titletext"

The simplest use is to place a text string in the key that spans the
entire key width rather than only the region of the usual title.

    plot A, keyentry "left-justified text" left, B

The new mode is also able to create a key with two columns of text,
by provided both sample text and title text in the same keyentry.

The sample text is left/right/center justified within the space allocated
for a key sample, so its position is affected by "set key samplen <width>"
The usual label formatting options, e.g. left/right/center/boxed,
may be appended for application to the sample text.

Example using dumb terminal:
    set key reverse Left samplen 2 title "Normal Key Title"
    plot f(x), \
    keyentry "A" left title "keyentry", \
    keyentry "B" center title "keyentry", \
    keyentry "C" right title "keyentry"

  +*
  | ***                                          Normal Key Title
  |    **
  |      ***                                      **** f(x)
  |         ***                                   A    keyentry
  |           ***                                   B  keyentry
  |              ***                                 C keyentry
  |                ***
  +---------------------------------------------------------------+