new command "remultiplot"
Remultiplot
-----------

Commands between "set multiplot" and "unset multiplot" are stored into
an internal datablock $GPVAL_LAST_MULTIPLOT.  They can be reexectuted
later using the "remultiplot" command, which is essentially a wrapper for
"load $GPVAL_LAST_MULTIPLOT".

This provides a partial answer to the long-standing request for a way to
replot an entire multiplot rather than just the most recent component plot.
A next step may extend this so that "remultiplot" is automatically executed
instead of "replot" or "refresh" when appropriate.

Note that the commands can be saved to a file by using
"set print FOO; print $GPVAL_LAST_MULTIPLOT".
In practice it may be best to save the pre-multiplot graphics context first
and then append the multiplot commands to it.

Known limitations
-----------------

The current graphics context at the time of "remultiplot" may not be the
same as the context at the time the multiplot was originally generated.
On the one hand this may be exactly the reason the replay is wanted.
On the other hand this can complicate reproduction of the original plot,
particularly if the commands in the multiplot sequence themselves change
the graphics context.

"pause" commands are ignored during multiplot replay.  I am not certain
exactly what goes wrong, but without this "remultiplot" fails to complete
and the program state is left part way through the multiplot.