conditional code to treat UTF-8 nonbreaking space as a blank
If in effect, this prevents leading nbsp from being misinterpreted
as part of a variable name.  The motivation was cut-and-paste from
from a web page using markdown to preserve formatting by rendering
leading whitespace as nbsp.  Gnuplot responded to each line with
"undefined variable:  "

If we decide to enable this check, a similar one should probably
be added to ALLOWED_8BIT_VAR().