allow iteration over array content
The iterator looks like this:
	array A = [1, 2, 4, 8]
	do for [a in A] { print a }

Thus it looks just like an iteration over words in a string,
but instead it uses successive entries in an array.

No checks are made for the type of the array entries.
You will, however, get the usual error messages if you try to apply
arithmetic operations to a string value or vice versa inside the
iteration loop.