Do not attempt to autoscale or extend axis ranges while zooming
If you zoom or pan an autoscaled data plot, points may go out of
range and autoscaling fails inside the following draw operation.
Usually this just means the zoom fails, perhaps with an error message
on the console.  Unfortunately this becomes a fatal error for the
wxt terminal if it happens inside event processing.

Even aside from the possible fatal error, it is unexpected to have
the plot view collapse to the remaining points rather than simply
representing the requested zoom box.

Now we test for in-progess zoom operations (this includes panning)
and use the requested zoom box y and y2 ranges without trying to
autoscale or auto-extend the axis.  We do, however, ensure that
the zoom box has non-zero width and height.

This does not fix the broken wxt error handling, but it does
prevent generation of y-axis range errors during pan/zoom operations.
It also makes the zoom behavior less surprising IMHO.

For 3D plots the requested zoom box is used for all three axes (x, y, z).

Related bugs: #2679 #2680