If libgd is not configured, fall back to libstbi for reading image files
Several plotting modes depend on being able to read in image data.
We use gdlib routines to read PNG JPEG and GIF files, so if gdlib
is not available then we lose support for
   plot ... binary filetype={png|jpeg|gif} with images
and for pixmaps.

This functionality can be replaced by a call to the public domain / BSD
image loader stb_image.  Documentation and source:
    https://github.com/nothings/stb/blob/master/stb_image.h
This could either be included as a source file or loaded as a library.

- modify the configure script to look for libstbi if libgd has not
  been selected or is not found

- modify the breaders routines with conditional code to call into
  libstbi rathern than libgd

- "show version long" indicates +LIBSTBI in this case

Notes

- without libgd we also lose the sixelgd terminal, but the kitty
  terminals are a reasonable alternative.