Reserve point types 1001 - 1020 for special use
Terminals that support polygon-fill can use this to draw
points with background fill.
Point types in the range 1001 - 1020 are reserved for this purpose.
This is obviously terminal-dependent.

As of this commit the extra point types are implemented for the
postscript terminal.
      1001    background filled square
      1002    background filled circle
      1003    background filled triangle
      1004    background filled upside-down triangle
      1005    background filled diamond
      1006    background filled pentagon

For the postscript terminal these map to the existing point types 70-75.
Note: This changes postscript point types 70-75 to use the current color
      for the shape border;  up till now it was always black with white fill.