qt: extra point types 1001 - 1006
Pass background color through to each new Point or set of Points

This allows the painter for a point to use background fill.

I dislike this approach since it adds overhead to every object created
but the only objects that care at this point are special point types
1001-1006.  It would be cleaner if any object that cares could retrieve
the background from the parent (root?) window/widget only when needed.