Enable the use of 'lc rgb variable' in 'splot with polygons'
Version 6 is moving toward a distinction between linecolor (used for the
border) and fillcolor (used for the interior) properties for filled-area
objects, including polygons.  It is not there yet but in the meantime it
is better to accept "lc rgb variable" as if it were "fc rgb variable"
rather than discarding the specified color and using black instead.

EAM: Modify the documentation to match.