New object type and plotting style "mark"
This patch implements a new plotting style called marks for drawing shapes
composed of polygons and polylines. It also suggests introducing another
style named linesmarks, which combines marks and lines styles.
Additionally, it proposes a new object named mark to place the individual mark.

In summary, this patch introduces the following new styles and commands:
• Plotting style marks (’plot ...  with marks’)
• Plotting style linesmarks (’plot ...  with linesmarks’)
• Object mark (’set object mark’)
• Command for defining a mark (’set mark’)
• Command for removing mark definitions (’unset mark’)
• Command for displaying mark definitions (’show mark’)

The shape of the mark can be a composite of polygons (or polylines).

The definition of mark data is established through the new command ’set mark’,
with these data managed using tag numbers greater than or equal to 0.

When employing the ’plot ... with marks’’ command, marks are drawn on the graph
with the specified coordinate series as the origin of each mark. Moreover,
mark shapes can undergo transformations such as translation, aspect ratio
deformation, and rotation at drawing.
The marks style primarily focuses on placing polygons, while the polygons
style is geared towards drawing polygons.

The purpose of introducing the marks style is to achieve terminal-independent
point marks with controling over fill style. Furthermore, it is to provide
convenience when drawing multiple marks of the same shape, as in tiling.
By limiting the mark components to lines and polygons, the data that gives
the mark definition is very simple, similar to the polygons style. These
features, while requiring some complexity in the preparation of mark data,
will be useful to users who are particular about the appearance of their plots.