Revise meaning of "mode" parameter
mode(0) FILLSTYLE
mode(1) STROKE ONLY
mode(2) FILL ONLY
mode(3) FILL AND STROKE
The default value has also been changed to mode=0 instead of mode=3.

Users who do not want to actively use mode can use fillstyle by default.
This means that mode can be omitted in the input sequence to set mark,
so it is sufficient to give only the (x,y) coordinates when defining the mark.

In the current implementation, a mark contains polylines and polygon's borders,
both of which are considered as 'stroke' and have the same color and thickness.