EAM - Make "ps N" and "ps var" act as they do for other plots
A pointsize given in the plot command (or from the global default)
is used as a multiplier for whatever size the mark would otherwise
use.  I.e. it applies on top of all the "units [xx|xy|...]" settings.

These all have the same effect:
	plot DATA using 1:2       with marks mk 2 ps 3
	plot DATA using 1:2:(3)   with marks mk 2
	plot DATA using 1:2:(1.5) with marks mk 2 ps 2
	plot DATA using 1:2:(3)   with marks mk 2 ps variable

These also have the same effect:
	plot DATA using 1:2:(45):(45) with marks mk 2 units xy
	plot DATA using 1:2:(90):(90):(0.5) with marks mk 2 ps variable

Caveat:
	I have only been looking at "plot with marks".
	I'm not sure if or how anything in this patchset changes
	the behavior of marks referenced by "set object".