marks 5 - reorganize struct marks_opts layout
This patch changes internal handling of 'marktype variable'.
Previously, it was determined by whether 'marks_options.tag'
was negative or not. Now, a member 'variable' to distinguish
whether the marktype is 'variable' or not has been introduced
to the struct 'marks_opt'.

In addition, the 'size', 'number' and 'interval' members,
which are no longer used in marks_opts, have been removed.