save_fillstyle did not handle default fillstyle with explicit border
Maybe there was no way to end up with this combination before,
but now you can hit it as in the example
    set mark N <foo> fs border lc "blue"

The difference from, e.g. "set obj 2 circle fs border lc "blue"
is that "set obj" resolves the default fillstyle immediately,
while the mark default is not resolved until it is used.