Should "plot for [i=N:*] ..." terminate if all points are out of range?
Currently the iteration terminates when there are no INRANGE points.
Bug #2712 points out that after "set clip two" this criterion fails
to plot graphs where the points are out of range but their connecting
lines as in range.

This commit tests for (n_outrange > 1 && clip_lines2) and keeps the
iteration going in this case after issuing a warning.

- Should this always be the case even if "set clip two" is not selected?
- Should it even bother to warn?