Remove ancient conditional code for SVR3 select()
In the early days (pre-1998) gnuplot acquired conditional code in
the x11 terminal driver on both the gnuplot and gnuplot_x11 ends
to work around incompatibility with the select() call in SVR3.
So far as I know we've been carrying it, unused, ever since.
Surely this is 20, if not 30, years past being useful to anyone.