Update pkg.m4 to match the one from pkgconf 1.8.0
Update pkg.m4 to match the one from pkgconf 1.8.0, which is an
API compatible [re]implementation of pkg-config.
The m4 file should work for either.

We modify this file to provide an alternative macro
PKG_CHECK_MODULES_NOFAIL so that the gnuplot configure script can
warn about missing modules but continue rather than failing with an error.