protect cexp(z) from underflow
cexp(z) underflows when creal(z) < E_MINEXP
This was previously being caught as a floating point exception by
fetestexcept(FE_UNDERFLOW),  but apparently something has changed
in glibc and cexp now sets errno=ERANGE and returns NaN instead of
zero and does not set a floating point exception.  This is annoying.

So now we have to test for creal(z) < E_MINEXP as a special case
and reset both errno and the return value to zero.

The change was easily visible in the 2nd plot of igamma.dem
where the upper surface disappeared (NaN) rather than reaching
a plateau value of 1.