generalize check for cexp(z) underflow into new routine gp_cexp(z)
The previous commit caught underflow only for Igamma(a,z).
Now we generalize this into a separate routine gp_cexp(z) and
replace cexp() with gp_cexp() in LambertW as well.

Other call sites for cexp() remain, but I am not certain that they
are prone to underflow and, if they are, what is the appropriate
treatment.