Better placement of axis tics on logscale axes (3D)
3D plots can use the same sanity checks for logscale tic generation as 2D
since the work is all done in the axis code.  However this revealed a quirk
that 3D plots (unlike 2D plots) may set up the axes for the initial plot
prior to terminal initialisation.  This required an additional check in
gen_tics() to allow the new initial tic-placement stage to succeed.

NOTES
- This changes the axis labelling in the one logscale plot in surface1.dem

Bug #2717