Allow "plot ... linestyle variable"
In very old gnuplot versions there was no way to change properties of the
default linetypes.  "linetype 1" was always red, had linewidth 1.0 and
associated point shape +.  So in gnuplot 3.7 if you wanted to change the
properties for lines 1, 2, ... you had to use a temporary "linestyle" instead.

The flexibility of assigning properties to line types has greatly increased
since then.  This has gradually made the use of line styles (as opposed to
linetypes) unnecessary, and many program features and plotting styles
introduced since then assume that you are using the new more flexible
linetypes rather than the older "linestyle" variants unless you explicitly
say otherwise.  So "linetype 1" is OK, and "linestyle 1" is OK,
but what the user might have intended for the color of the second plot,
or the second variable color became ambiguous.

Gnuplot version 4.2 (2007) introduced the command
        set style increment userstyles
that told the program to use linestyle N in preference to linetype N
if both were defined, and for successive clauses in a "plot" or "splot".

Gnuplot 5 consolidated most syntax using a numerical line to default to
the linetype unless "set style increment userstyles" was in effect.
So for example "plot ... with lines lc variable" interpreted input
integers either as linestyles or linetypes depending on "set style incr".
However not every relevant place in the code checked for this.

Gnuplot 6 has deprecated "set style incr" and requires the user to
distinguish between "ls N" and "lt N" everywhere.  The program does not
try to double-guess which is wanted.

However, the initial release of 6.0 failed to handle the case
    plot ... linestyle variable
so interpreting integers in a data file as linestyles was difficult.

This commit makes "[ls|linestyle] variable" valid syntax for a
color specifier.

Bug 2719