modify (hopefully improve) parsing of the "help" command
I got tired of dialogs like this:
    gnuplot> help watchpoint
    Ambiguous request 'watchpoint'; possible matches:
	    watchpoint labels
	    watchpoints
    gnuplot> help watchpoint label
    Ambiguous request 'watchpoint'; possible matches:
	    watchpoint labels
	    watchpoints
    gnuplot> help watchpoint labels
    Ambiguous request 'watchpoint'; possible matches:
	    watchpoint labels
	    watchpoints

A perfect match to the first word of the help request should not
be considered ambiguous and should not be expanded to a longer
word belonging to a different help entry.

Caveat:  This code is really tangled.  I am not 100% the modified
code is correct either, although it does resolve the issue shown above.