wxt: handle mousing of logscale axes in inactive plot window
When mousing an inactive plot window, the axis limits and scaling
are retrieved from a per-plot structure initially sent by wxt.trm
and stored wxt_axis_state[].  However only the limits and scaling for
the user-visible axis were sent, not the underlying primary axis.
This was a oversight from all the way back to version 5.

In the case of logscale this is easy to fix - we just send the
primary axis info instead and let wxt_gui do the conversion
user_coord = primary_coord ** logbase.

In the case of other nonlinear or linked axes the necessary
"via" and "inv" functions are not available to wxt_gui.
For now this results in the coordinate being reported as NaN.

Bug 2723