Implement "splot with filledcurves" as a varant of zerrorfill
The primary motivation is to make it easy to create waterfall plots.

    set style fill solid border lc "black"
    splot for [scan=N:1:-1] DATA index scan \
          using 1:2:3 with filledcurves fc background

Apart from being easier to explain than the 5-column zerror equivalent
this has the advantage of drawing only the curves, not a closed polygon
that connects the endpoints of each curve.

Note: I do not know why these are called "waterfall" plots.
I am familiar with them as a way to show the evolution of a signal
over time, or to compare spectra recorded sequentially as something
changes.  They are similar to "fence" plots.