New command "save changes <filename>"
This is an internal implementation of Petr Mikulík's gpsavediff
utility script.  The current output from "save" is compared to
the output from "save" run at the start of the session.
Only the changes are saved to the output file.
Output to a pipe is OK.

The two save files are opened by fp=tmpfile() and subsequently
accessed as "/dev/fd/<fileno(fp)>".
Their content is piped through /usr/bin/diff and /usr/bin/sed
to generate the output stream.

The details of this method make it not suitable for Windows or OS2.
I have not tested on MacOS or any other non-linux systems.
Please report success or failure.

Example:
    save datablocks "myplot.gp"
    save changes "myplot.gp" append
    ...
    reset session
    load "myplot.gp"

Advantages:
The full "save" command produces about 200 lines of output and it can
be very hard to determine which lines are specifically relevant to
reproducing the final plot.  The output from "save changes" is much
shorter and thus easier to understand and edit if necessary.

Disadvantages:
Since only the changes relative to the initial session state are saved,
loading the saved file without resetting the session may not duplicate
the saved plot.  If you want to restore the full context independent of
the current session state, the complete output from "save" is better.