docs and demo for support of Delaunay triangulation
Delaunay triangulation was added as a filter operation used as a
first step in generating concave hulls.  However the filter itself
has not been documented as a separate operation available to the user.

Modify the demo file "delaunay.dem" in the demo collection and refer
to it as an example in the user manual.

Still marked EXPERIMENTAL and depends on ./configure --enable-chi-shapes