fixes for various problems seen for "splot with circles"
It would be nice if all the 3D plot styles with fill properties
would use the same conventions for where to store the colors
and what field + flag is used to indicate variable color.
But they don't.
So these fixes address only problems for "splot with circles".
Some of these issues are only partially fixed.

- support for per-plot fill style
- variable fill color inconsistently applied  (fc variable, fc rgb variable)
- "lc <colorspec>" is sometimes required even though "fc <colorspec>" is intended
- key sample does not always match the plot
- "fs solid border lc variable" is probably hopeless