remove dead code and comments related to DISABLE_SPACE_RAISES_CONSOLE
The configuration option --disable-raise-console is relevant only
to Win32 and to the x11 and qt terminals.  Furthermore it is only
relevant to qt under the (long gone) KDE3 desktop or its cloned
offspring Trinity.

Remove all the comments elsewhere that say "if we supported it,
do something here".
I am tempted to remove the configuration option altogether since
a Windows build doesn't use ./configure anyhow, but for now I leave
it in place even though only x11 or qt-under-Trinity would notice.