Rework term->waitforinput() to accommodate UTF-8 with BSD editline
The editline library apparently does not check that stdin is really
editable, so we must avoid calling el_wgetc() if we are taking
input from a pipe rather than a terminal.
Note:
    TBOOLEAN interactive
is basically set to isatty(fileno(stdin)) on systems that support it.
It might be better (more foolproof) to test directly against isatty()
but that might not work when building for WIN32 or WGP_CONSOLE.