tikz: The "nostandalone" option did the opposite of its intent
Both "standalone" and "nostandalone" caused the tikz terminal to a
minimal wrapping document for the plot(s).
If neither option was given, no wrapping document was emitted
(clearly what "nostandalone" was intended to do).
	set term tikz			no wrapper
	set term tikz standalone	wrapper
	set term tikz nostandalone	wrapper (bug!)

The "fulldoc" and "nofulldoc" synonyms for "standalone" and "nostandalone"
suffered the same bug and get the same fix.

gnuplot-tikz.lua script version is bumped to 120

Bug  #2740