metapost:  updated to support template for output file names
	set term mp template "outputtemplate"
adds a line to the gnuplot output file that will be fed to Metapost.
	outputtemplate:="outputtemplate";

The effect is that Metapost will use this template to create file names
for the epsf files it creates from plot commands in the *.mp file.
"%j.mps" will yield a single file with extension `.mps` instead of `.0`.
    set output "myplot"
    set term mp template "%j_%c.mps"
will yield files myplot_0.mps, myplot_1.mps, etc.

Make sure to include "%c" in the template when multiple output files
are generated.  Not specifying a template is equivalent to specifying
outputtemplate="%j.%c".