Handle a corner case of intersecting pm3d surfaces
In general when two pm3d surfaces intersect, any given quadrangle in
the surface either has no intersection or one line of intersection with
the corresponding quadrangle of the other surface.  However there is a
pathological case in which the quadrangle wraps around a ridge of one
surface such that alternate corners of the quadrangle are above and below
the other surface and thus there are two lines of intersection within
a single quadrangle.  This case was not handled previously.

Bug #2744